- Comparable<> интерфейс - Comparator<> интерфейс
